biocentral_server/pyproject.toml
[project]
name = "biocentral-server"
version = "0.2.5"
description = "Compute server for the biocentral frontend"
authors = [
    { name = "Sebastian Franz", email = "sebastian.franz@tum.de" },
    { name = "Tobias Olenyi", email = "tobias.olenyi@tum.de" },
]
license = { text = "GPL-3.0" }
readme = "README.md"
requires-python = ">=3.11.9,<3.13"
dependencies = [
    # Backend
    "taxoniq>=1.0.1,<2.0.0",
    "blosc2>=2.7.1,<3.0.0",
    "psycopg[binary]>=3.2.3,<4.0.0",
    "gunicorn>=23.0.0,<24.0.0",
    "rq>=2.1.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "gpytorch>=1.14,<2.0",
    "onnxruntime-gpu>=1.16.0,<2.0.0; sys_platform == 'linux'",
    "onnxruntime>=1.16.0,<2.0.0; sys_platform != 'linux'",
    "tritonclient[grpc]>=2.61.0",
    "circuitbreaker>=2.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "fastapi>=0.120.2",
    "uvicorn>=0.38.0",
    # Git Dependencies
    "biotrainer @ git+https://github.com/biocentral/biotrainer.git@fix/api-improvements",
    "protspace @ git+https://github.com/biocentral/protspace.git@chore/matplotlib-optional",
    "hvi_toolkit @ git+https://github.com/SebieF/hvi_toolkit.git@maintenance/update-biotrainer",
    "vespag @ git+https://github.com/biocentral/VespaG.git@onnx-deployment",
    "tmbed @ git+https://github.com/biocentral/TMbed.git@onnx-deployment",
]

[dependency-groups]
dev = [
    "pytest==8.3.1",
    "pip-audit>=2.7.3,<3.0.0",
    "ruff>=0.12.1",
    "pre-commit>=4.2.0",
    "pytest-asyncio>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"


[tool.uv]
override-dependencies = [
    "perf-analyzer>=2.59.1; sys_platform == 'linux'",
]
